Refactor excel
